Rename file and add typed BaseStore
